Split Cloud Functions into separate files
