+++ functions/export-csv.js
@@ -1,0 +1,33 @@
+/**
+ * @license
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+const {db} = require('./common/context');
+
+function exportCsv(req, res) {
+  const {
+    project: providedProjectId,
+    featureType: providedFeatureTypeId,
+    lang: desiredLanguage,
+  } = req.query;
+
+  return res.status(500).send('Not supported yet.');
+}
+
+
+module.exports = {exportCsv};
