+++ functions/on-create-record.js
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict';
+
+const {db} = require('./common/context');
+const {getSheet} = require('./common/sheet-config');
+
+function onCreateRecord(change, context) {
+  const {
+    projectId,
+    featureId,
+    recordId
+  } = context.params;
+  const record = change.data();
+  const {
+    featureTypeId,
+    formId
+  } = record;
+  return getSheet(projectId).then(sheet =>
+    sheet && sheet.getColumnIds().then(colIds =>
+      db.fetchFeature(projectId, featureId).then(feature =>
+        sheet.addRow(feature, featureId, recordId, record, colIds))));
+}
+
+module.exports = {onCreateRecord};
